Update README.md
Objective of this report is to apply a credit risk model analysis to create a credit scorecard. 
The scorecard will use for decision making and evaluating loan applicants based on the 
provided loan historical information. The methodology used for this scorecard is binning with 
WoE transformation and a logistic regression model.  This report will guide readers to 
understand the analytical process and steps.